+++ webview/src/pages/context.ts
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const WebviewPageContext = React.createContext<{ refreshDocFlag?: boolean }>({});