+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "tswagger" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+### Changed
+
+- Unnecessary to rename `pathParamName`.
+
 ## [1.2.6] - 2024-01-10
 
 ### Changed