+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "tswagger" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+### Fixed
+
+- Duplicate results were generated after refresh.
+
 ## [1.2.7] - 2024-01-11
 
 ### Changed