+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "tswagger" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+### Changed
+
+- Compatible with more situations of the api definition.
+
 ## [1.2.8] - 2024-01-14
 
 ### Fixed